Fixed warnings on MacOS port.

+++ sources/Core/IndexFormat.cpp
@@ -14,7 +14,7 @@ namespace LLGL
 
 
 IndexFormat::IndexFormat(const DataType dataType) :
-    formatSize_( DataTypeSize(dataType) )
+    formatSize_( static_cast<unsigned int>(DataTypeSize(dataType)) )
 {
 }
